+++ bootique-linkmove2-rest/src/main/java/io/bootique/linkmove/v2/rest/LinkMoveRestModule.java
@@ -21,7 +21,9 @@ package io.bootique.linkmove.v2.rest;
 import io.bootique.ConfigModule;
 
 /**
- * @since 2.0.B1
+ * @since 2.0
+ * @deprecated in favor of LinkMove v3
  */
+@Deprecated(since = "3.0", forRemoval = true)
 public class LinkMoveRestModule extends ConfigModule {
 }