+++ bootique-linkmove2-rest/src/main/java/io/bootique/linkmove/v2/rest/RestConnector.java
@@ -9,8 +9,10 @@ import java.io.InputStream;
 import java.util.Map;
 
 /**
- * @since 2.0.B1
+ * @since 2.0
+ * @deprecated in favor of LinkMove v3
  */
+@Deprecated(since = "3.0", forRemoval = true)
 public class RestConnector implements StreamConnector {
 
     static final String TEMPLATE_VALUE_MAP_PARAMETER = RestConnector.class.getSimpleName() + ".templateValueMap";