+++ bootique-linkmove2/src/main/java/io/bootique/linkmove/v2/LinkMoveFactory.java
@@ -46,8 +46,10 @@ import java.util.Optional;
 import java.util.Set;
 
 /**
- * @since 2.0.B1
+ * @since 2.0
+ * @deprecated in favor of LinkMove v3
  */
+@Deprecated(since = "3.0", forRemoval = true)
 @BQConfig
 public class LinkMoveFactory {