+++ bootique-linkmove2/src/main/java/io/bootique/linkmove/v2/LinkMoveModule.java
@@ -31,8 +31,10 @@ import javax.inject.Singleton;
 import java.util.Set;
 
 /**
- * @since 2.0.B1
+ * @since 2.0
+ * @deprecated in favor of LinkMove v3
  */
+@Deprecated(since = "3.0", forRemoval = true)
 public class LinkMoveModule extends ConfigModule {
 
     /**