+++ bootique-linkmove2/src/main/java/io/bootique/linkmove/v2/LinkMoveModuleProvider.java
@@ -30,7 +30,9 @@ import static java.util.Arrays.asList;
 
 /**
  * @since 2.0
+ * @deprecated in favor of LinkMove v3
  */
+@Deprecated(since = "3.0", forRemoval = true)
 public class LinkMoveModuleProvider implements BQModuleProvider {
 
     @Override