+++ bootique-linkmove2/src/main/java/io/bootique/linkmove/v2/connector/IConnectorFactoryFactory.java
@@ -27,8 +27,10 @@ import io.bootique.config.PolymorphicConfiguration;
 import io.bootique.di.Injector;
 
 /**
- * @since 2.0.B1
+ * @since 2.0
+ * @deprecated in favor of LinkMove v3
  */
+@Deprecated(since = "3.0", forRemoval = true)
 @BQConfig
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
 public interface IConnectorFactoryFactory<C extends Connector> extends PolymorphicConfiguration {