+++ bootique-linkmove2/src/main/java/io/bootique/linkmove/v2/connector/JdbcConnectorFactory.java
@@ -28,8 +28,10 @@ import io.bootique.jdbc.DataSourceFactory;
 import javax.sql.DataSource;
 
 /**
- * @since 2.0.B1
+ * @since 2.0
+ * @deprecated in favor of LinkMove v3
  */
+@Deprecated(since = "3.0", forRemoval = true)
 public class JdbcConnectorFactory implements IConnectorFactory<JdbcConnector> {
 
 	private DataSourceFactory dataSourceFactory;