+++ bootique-linkmove2/src/main/java/io/bootique/linkmove/v2/connector/JdbcConnectorFactoryFactory.java
@@ -27,8 +27,10 @@ import io.bootique.di.Injector;
 import io.bootique.jdbc.DataSourceFactory;
 
 /**
- * @since 2.0.B1
+ * @since 2.0
+ * @deprecated in favor of LinkMove v3
  */
+@Deprecated(since = "3.0", forRemoval = true)
 @BQConfig
 @JsonTypeName("jdbc")
 public class JdbcConnectorFactoryFactory implements IConnectorFactoryFactory<JdbcConnector> {