+++ bootique-linkmove2/src/main/java/io/bootique/linkmove/v2/connector/URIConnectorFactoryFactory.java
@@ -35,8 +35,10 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * @since 2.0.B1
+ * @since 2.0
+ * @deprecated in favor of LinkMove v3
  */
+@Deprecated(since = "3.0", forRemoval = true)
 @BQConfig
 @JsonTypeName("uri")
 public class URIConnectorFactoryFactory implements IConnectorFactoryFactory<StreamConnector> {