+++ bootique-linkmove2/src/main/java/io/bootique/linkmove/v2/resource/BQUrlResourceResolver.java
@@ -28,8 +28,10 @@ import java.io.Reader;
 import java.net.URL;
 
 /**
- * @since 2.0.B1
+ * @since 2.0
+ * @deprecated in favor of LinkMove v3
  */
+@Deprecated(since = "3.0", forRemoval = true)
 public class BQUrlResourceResolver implements ResourceResolver {
 
     private FolderResourceFactory baseFolder;